Importing history: Snapshot 2009/09/27 01:48
The progress of the project was rocky after INT02 and not much effectively
got done.

We do now have a second module, though and the folder structure is already
a lot like what it got to be. However, the only real change since INT01 is
work on GoodManners, which got reworked several times and was split into
several modules before it finally got into a working state.

It's also interesting that we already have an empty folder for GoodSport
here (though git won't preserve it), which is basically the module that
held up the GoodLooking development for so long and which is at the time
of writing (years later) still unstarted (though it's one of the next
things on my todo-list, so it should get to be not too far into the
future.