Importing history: Snapshot 2011/12/21 13:25
This looks like it was the first snapshot after I "saw the light". It is
only two days newer than the last snapshot, but that may not mean much
(it's very possible I was thinking about it for longer already, started
coding two days before this and just decided to do a snapshot before
touching my codebase).

Anyway, we now get the GoodRoleModel module and with it our datatype
format to create our data definitions in. This looks a lot like it was
made with GoodService in mind, but none of it is there yet. GoodManners
still also isn't yet split into GoodManners and GoodMemory, but things are
definitely moving in the right direction.

I don't believe much has changed in GoodRolemodel between the moment of
this snapshot and comitting (which is with a working chain of GoodMemory,
GoodManners, GoodService and GoodRolemodel.