Importing history: Snapshot 2012/10/17 23:03
Once again, we have to wait about a year before our next snapshot.
However, this snapshot brings so many new things to the table that I doubt
that I didn't work on the project during that year.

Perhaps the smallest of the changes is one I do quite like: each model now
has a description. The description contained how the module got to be,
sometimes a bit more history and how I came up with its name. If you
can't like this for any other reason, like it for the fact that it makes
sure that GoodSport is no longer an empty folder, so it is finally
recognized by git.

GoodRolemodel sees some small changes, while GoodService is written from
the ground up. So is GoodMemory (though it's not completely done yet) and
basically GoodManners is too, as the previous design was reworked
compeletely.

I ended up temporarily shelving the project because of time constraints.
The driving force behind this explosion of improvements to the code had
been the fact that I wanted to use it for a rewrite for a system that I
was going to need shortly. After spending quite a bit of time on this
project, I realized that a rewrite wasn't actually feasible (it may not
even have been feasible if the Good Suite had been done at the moment in
time I started working on the suite). As such, I shelved the project while
I was working on GoodMemory and quickly added some features for the
fixed the bugs in the application (I normally used the application once a
year, but I had skipped the year before and in those two years the web had
changed quite a bit so there was quite a bit to fix) and I added
functionality based on the feedback from two years before that. I also
promised myself I was going to do a rewrite before needing it again the
year after and I decided that I would be able to do it using the Good
Suite simply because it was in a workable state by that time.