Importing history: Snapshot 2012/11/13 01:55
We are finally getting to a point where the snapshots are close together
in time. The changes in this snapshot were mostly based around getting
quad of GoodRolemodel/GoodService/GoodManners/GoodMemory in working order.

For that to happen, the test were moved out of the main directory (and
being manual tests, I'm not even putting them in git anymore at all), it
means that GoodMemory was now finished form the broken state that it was
lef tin before and it meant that a lot of code was actually tested for the
first time, which explains why there are small changes all over the code.

In addition to this, the way the Storables as created by the
GoodServiceCompiler when under influence of the GoodMannersStorable
communicated with the GoodMannersStore (and in Particular the
GoodMemorySQLStore), which meant the /Good/Manners/Report folder
disappeared and meant that GoodMannersStore and GoodMemorySQLStore now
became compiled classes as well.