Importing history: Snapshot 2012/12/18 03:05
The Resolver classes.

As we got closer to the moment we started using this repository, the
snapshots became much more like commits. As such, this snapshot only
introduced one feature: The Resolver classes.

These classes changed the way in which it was determined how deep to
resolve the classes when getting them from the database. This cleaned up
the interface, but it also allowed for the same classes to be used in
order to allow specifying a way to order the results of the "query". This
was the real reason this change was brought about.