Adds Dummy Storables
A dummy storable is an instance of the object that pretends to be fully
functioning, but in reality has only an id that was given to it on
creation. It is for referring to objects when the id was acquired at
another time and the id was stored.

It still has to be made so that the dummy gives an error when trying to
access anything other than its id. Additionally, it might be nice if it
was possible to resolve a dummy and make it a full object.

In addition to creating the Dummy, two more changes had to be made to make
them work. First, it was not a new object, so it was failing the check if
it was either new or from the current Store. This was removed by checking
if it either has no store or has the correct store. The idea is that that
will have be redone anyway, so not too much attention was given to it. The
other change was about allowing the dummy to be used to get an object by
id, which required us to check if the object has a valid id when a
comparison condition comes up empty.