Adds LookingWithManners for module interoperation.
A better way to be able to use GoodLooking with GoodMannersStorables needs
to be made, but until the time that there is such a method, this will have
to do.

This simply converts GoodMannersStorableCollections into arrays where all
the result objects are members of an array and turns GoodMannersStorables
into associative arrays.