Fixes bug that caused Storables to be null.
When a property name of a property that was referring to another datatype
(or table, in the sql context) had capitals in it was silently being
turned into null when an object with that property was fetched from a
Store with that property resolved. This was because the field name in the
sql query wasn't being made lowercase.

This commit fixes that problem.