GoodLooking Rewrite
Closes #1

We're still far from a proper system, as the parser is actually only a
partial parser (parses up to statement level and leaves statements strings
to be parsed at compile time) and we're still using a messy regex based
parsing strategy instead of a proper lexer + parser combo.

However, at least the code has a decent structure now, is something that
resambles clear and can easily be extended or modified.