Fixes GoodLooking bug when last content is code
When the last part of a template was a inside code tags, the parser would
fail parsing it. This was a mistake in the way the input was consumed from
the input string after it was parsed.

This bug was found while writing the first tests for The Good Suite and
was fixed instead of reported.