Fixed parentheses as mathematical prioritzers
During parsing, the parentheses were correctly rezognized, but instead of
continueing with the input minus the parentheses, it continued with the
exact same input, causing an infinite loop.

This bug was found while writing the first tests for The Good Suite.
Really goes to show how important tests can be.