Prevents parsing reserved words as variables
It's a pretty ugly hack, but it's really hard to do this properly. On top
of that, the hack was already there for "else", it just needed to be
extended to include the "end <something>" (it got quite a bit uglier,
though.

The problem that led me to find this was that "end if" was recognized and
"endif" wasn't. Though I will be changing this some time soon anyway, this
was a bug and allowing both was the way it was intended.