Adds xor operator to GoodLooking
I wrote tests for it (yet to be commited) and implemented it under the
impression that it was described in syntax.txt. Afterwards, I found it
wasn't. However, it was a good excercise, little work and in the end there
is no good reason not to include it in the syntax.