Fixed outer join problem
Changed from inner joins to left outer joins. This makes sure that rows
with a null value in a reference field are also selected, which makes
sense with out API.

Yet another victory for testing, as during the during the writing of tests
lots of bugs are found and most of them are fixed.