Fixed bug with null fields and resolvers
There was a problem with null fields in a query result and creating the
object that went with that result.
The column count had to be updated when there was a null value, but not
when the column wasn't present, but it wasn't increased in either case.

To fix this bug, some cleanup was done in the object creation code in
SQLStore(Compiler). It's still messy, but not as messy as before.