Removed code that wasn't sanely reachable
I believe I added this code copied from one of the other classes in an
attempt to navigate the minefield of bugs. However, now that many of the
bugs have been taken out, it is clear what the code exactly meant and ti
just doesn't make sense.

Basically, it was code that was only reached if you provided a non-new
object as the root for a condition that was then used to do a modifyAny,
but that makes no sense to do and should just be replaced by a normal
modification in any case.