Changing names names and structure of RoleModel
A huge refactoring change, this commit introduces the new namespace
RoleModel\Schema, renames DataModel -> Schema, changes a couple of names
inside that namespace and folds the DataMember class into the Type
abstract class, renaming it and its children to Member{Type}.

Hopefully, this makes the naming a bit clearer, and the change in the
number of visitor function should also make less code necessary (now, we
have what we need when we need it, so we no longer need to store what we
need in a member variable until we need it).

Because this changes the type you need to pass to compilers and because
the Visitor interface was implemented in so many places, there are changes
all over the board.