Removes Store from Condition constructor
This was present to do a check that it wasn't checked against a store
other than the one it was created with. This check is also removed (as is
the requirement for it to be the same store). This was part of the same
overengineering as 901b79f3. Together with that commit, this should allow
some changes which make for a nicer API, as well as some new interesting
tricks with Conditions.