Removes createCondition functions from Store
And introduces the new API, in which the user can just create Conditions
him or herself. This is better because it allows users to create
Conditions without creating a Store first, it allows them to use a single
condition on more than one store (as long as it makes sense) and while the
names remain long, the use of namespaces gives control over how to write
it back to the user.

This was made possible by 035b4c0a and 901b79f3.