Cleans up Modifier\Storable (a bit)
Some unnecessary code was removed, including a number of un-used functions
(one of which was for user code, but other user code will soon be provided
in its place).

One of the functions was called from SQLStore, so the call has been
replaced by other code that does the same thing and does not need a
seperate function.