Removed the concept "fieldnumber" from SQL writers
This was used solely for the storing and retrieving of joins. It was
rather contrived because it counted all the reference field, but not the
non-reference fields.

Now, we just use the fieldname instead, it is just as unique, it is less
contrived and it can also be used in places where we do not know which
field is a reference and which is not.