Removed double field/tablenamifying on Join array
This was being fieldnamified and tablenamified twice: once during creation
of the Join object and once when inserting the values in the SQL
statement. This isn't a problem at the moment, but when something like
table prefixes is added, it will suddenly start causing bugs.

I decided to remove the ifying when creating the Join object, because it
felt right to keep the "escaping" as close to the SQL as possible. In
addition, this makes it possible to use the join data when actually
interested in the type name, not the table name (such as when building a
Storable from the sql results).