Makes Store and SQLStore non-generated.
There were some reasons for having these classes, but it was better not to
generate outside the datatype classes.

This means a change in the public interface, as now we no longer have
separate insert and modifyAny function for each class, but just one
generalized version, which will detect the type automatically.

The creation of the Resolver class was moved to Modifier\Storable which
also means that the extraFiles mechanism is taken back into use.

The entire system that built objects upon Collection::getNext() had to be
remade and part of that change led to FieldDenamiefier (which isn't as bad
as it sounds, it uses a visitor-loop to see if the fieldnamified fiels are
in the original and if they are, recreates them in the output with the
untouched key, it doesn't actually try to reverse the namifying algorithm.

This closes #7.