Implemented a "zero protected" policy for Storable
The rationale is that this means there's less for collisions between
library defined functions and user defined inherited, since children are
simply not aware of their parents' privates.

This means some code duplication, but that only means a file size increase
for the generated files, and that amount is negligible anyway. The usual
nightmares with maintaining the code do not apply, since the code is
generated, meaning it's only in one place in the library code anyway
(it has just moved down a number of lines within the same file).