More Grammar cleanup (+important bugfix)
Where the first round of cleanup was about the way the class fitted into
the rest of the classes and making the class look better, this second
round is about an actual rewrite of a part of the class.

Basically, the handling of control structures has been redone in such a
way that the entire grammar is now defined in this single class. There is
still some whitespace rules in other places, but other than that the
syntax of our control structures is now in the hands of the Grammar class.
This meant rewriting how some things were done and removing some rather
poor regexes from the class as they were no longer needed.

There is yet one more change to come.

Finally, this fixes #29 (which was still marked as "unfixable" (without a
major rewrite - far more major than I did here) even though I did realize
that it was fixable after all a while back.