Last Grammar cleanup
Over the process of our cleanups, we went from static function to numerous
instances of the class, which meant str_replace(ish) things being done
more often than necessary.

I've brought this back to just a single instance which a lot of classes
have a reference to. What we get as an advantage over the original
situation is that not only do we now just have cleaner code that is more
readable, it is now also possible to reuse large portions of the code with
a different Grammar.