For -> forrange
This is one of the few steps left for #19

I don't completely like forrange yet. The double r is misleading and the
forrange looks like it should be pronounced completely differently
(somewhere between orange and forage). However, I'll just stick with this
for now, and see if I like it enough or can come up with a better
alternative.