Adds protectedness to generated getters/setters
(when appropriate, only, of course)

This closes #62.

This tests for Exceptions, as it was the only test for the absence of
properties. Of course, this should be changed in #22 when the Excpetion
system gets its overhaul. (However, that should be obvious at that point
in time, as these tests will just fail after that change.)