Allows reuse of automatic vars in Good Looking
This came up while solving #68 and is now fixed. It is not allowed to
overwrite a variable with another type of variable and you can't reuse a
loop variable inside that loop (otherwise infinite loops would be
possible) but you can reuse a loop variable for another loop *non-nested*
loop.