Add Storage to Id constructor
Id shouldn't be seen as "some object with id 'x'", but rather as "a
representation of the object with id 'a'". Add to that that ids are only
unique per Storage (not the instance, but the underlying storage, such as
the sql database) and it makes no sense to talk about an id without
linking it to a Storage.

The store isn't checked when used in a coondition, or when set on an
object (either a real object or a modification object) but semantically
that still only makes sense in the context of a store, since otherwise you
won't know what the id is going to refer to.

This will make the Id interface for get() and delete() cleaner, and as
such, we're making our way to #76.