Fix tests that used old API
One of the tests was still using a pre-#49 API, which accidentally still
sort of worked. (Previously, it used an associative array to link datatype
names to the files, now it expects just an array with files, and if that
array is associative it will still work just the same. However, it just
discards the key information (which previously had a meaning), so it is
just wrong.

Another test copied this incorrect behaviour, so it had to be fixed as
well.