Fix bug in Rolemodel (previously untested)
The type modifiers regex was matching the empty string (as well as just
whitespace) because the idea is to allow brackets with no arguments in the
datatype files. However, the processing code didn't take this into
account.

The best way to fixed this seemed to be to remove the case outside
capturing group, since that covers all corner cases without introducing a
lot of code or another regex (to match the specific whitespace case).