[#122] Allow ints in float properties
Unlike most changes, this is not a change in phpunit, but a change in
the cpde that this is testing.

Int values are now allowed in float properties, so that test should
throw no exception. Additionally, I added a test with the original
intent of that test (an invalid value in a float property): by setting a
string value now.