[#122] Rename type to prevent name collision
Excluding case (as php does) the type had the same name as one of the
methods that was generated on it. The collision part is in that php now
considers that method to be the constructor.

Interestingly, this may have been present previously. However, what
brought it to light is that this is now *deprecated* (you should use
__contstruct instead).