[#129] Make compiler follow new Schema structure
The changes aren't quite as big for the Storable Modifier, but there
are still few lines that do not make the diff...

(The reason the change is smaller is because both the original code
and the new code need to follow the approximate structure of the
Schema, which is basically the same. For the Storable Modifier, it
wasn't because it was originally following the structure *twice*.)

There is also a small change: the compilation process no longer returns
an array of files it generated. I didn't use this feature anywhere
(except for in the tests) and it didn't really make sense to me either.