[#96] Refactor type checking system
I needed some sort of Type system in order to support writing of queries
(because there, I need to know the type of a Collection). I went with
the best solution I could come up with.

This solution presented the opportunity to turn the type checking system
into something nicer. Especially the less clean solution I recently made
for checking the type of an element added to a collection derived a lot
of benefit from this change.