[#96] Implement inserting of collections
The StorableCollectionEntry approach is not optimal in speed, in a
number of ways. But it does work, and that matters even more.

The StorableCollectionEntry also is a rather minimal implementation of
Storable, but I think that will improve as I get around to adding more
SQL functionality and find what methods are needed. I'm not planning to
implement methods that don't appear to be needed, though.