[#96] Implement simple updating of collections
This also replaces the insert system, as it was easier to cover that
scenario in this new solution than not to do so.

There is still a problem with detecting whether a change has been made
to the collection or not, which will cause changes not to be persisted
if the first change you make to a storable is to one of its collections.