[#96] Rename [Storable]CollectionEntry[Storable]
(So: "Storable" was in the front of the name and it's now in the back.)

This is a Collection Entry that is also Storable, not an Entry of a
Storable Collection or something like that. I think the new name is less
ambiguous about this, even if it is somewhat less natural.