[#96] Rename Storable-CollectionModifier+Storable
(So: "Storable" was at the beginning and is moved to the beginning, the
opposite of the last change.)

This is another name change sacrificing a little "naturalness" in
exchange for less ambiguity. This is not a Modifier of
StorableCollections. It is a modifier of collections, so these
collections can be used with Storables (and though the collections do
not get the interface themselves, this is a key piece in storing them in
a storage.)