[#143] Major rewrite of the condition system
This is a change of a public API! (but you know, 0.x.y)

The collection-based conditions didn't really fit properly in the
old Condition API. So, I just replaced the whole Condition API by a
better one. In a way, I think this is basically something that I always
preferred, but when I couldn't think of how to make this work, I ended
up with the previous system.