[#143] Clean up ConditionProcessor/Type mess
I had made a bit of a mess of this, so it was time to clean it up. Some
of it still isn't as good as I would want it to be, but the
Condition/Comparison situation plays into that, so I can't fix it until
I've also got that cleared up.