[#143] Re-use Selecter in for update conditions
So, I was trying to understand the code in the UpdateConditionWriter and
the SQL it generated. I found that the query was pretty complex and
could very easily be brought in a position where it used multiple
subqueries. In this process, I also found a situation in which it wrote
an invalid query.

Instead of trying to fix this, I decided to reuse the Selecter. There
now always a subquery, even if not required, but there will never be
more than one and the code complexity is way down because we now reuse
code elsewhere (which wasn't as complicated to begin with to boot).