[#145] Merge the Condition and Comparison classes
The two concepts aren't logically merged yet, so this is very much a
work in progress. However, the only one of the two interfaces now
exists, which is quite a bit change already. And besides, the whole
things passes the tests again, so it made sense to use this as a moment
to make a commit.

This does introduce the ComplexCondition interface without any methods.
I found this to be a manageable way to fix code using "instanceof" that
broke. I think I won't keep this interface around (but I'm not
completely sure) but the concept of ComplexCondition is one I think I
will still need in the future.