[#145] Get rid of the EqualityComparisonProcessor
I just replaced it with a bit that throws an Exception if an illegal
comparison is used for the context. There's a second part to this, which
is to check if the structure is valid when making it, but that's a later
step. I should get to that soon, though.