[#145] Make sense out of Collection Conditions
Ths actually stayed mostly the same. The new bit is actually untested
because it can't be used with the current API (which includes the
CollectionComparisonSHolder) but I plan to change that (removing that
(CollectionComparisonsHolder) next, so I should test it at that point.

The way that the AndCondition is a Condition and a CollectionCondition
but the two aren't related is far from perfect, but it works...