[#145] Remove CollectionComparisonsHolder
This makes the CollectionCondition API more in line with the rest, and
makes it possible to use either AND or OR on collections.

There were tests for AND that became impossible, so I changed them to be
changed to the new way and added the OR tests. This meant code that I
was untested was now tested and it needed quite a few changes before it
worked properly too.