[#145] Remove Comparison remnants from Processor
(To be specific, from the ComplexConditionVisitor)

I'm not entirely happy with the way I now use the
ComplexConditionDiscoverer, but it works (and it proves that no other
methods are needed.

I also combined the processReferenceMember and processPrimitiveMember into
processMember because they were basically the same. Maybe I will one day
be able to get the processCollectionMember in there as well, but I
wouldn't want to give up the fact that there's type hinting (and
assurances) and with Condition and CollectionCondition completely
separate right now, that's not possible at the moment.