[#145] Remove Comparison remnants from Storables
The generated classes still had two fields: one for comparisons and one
for conditions. Since both are the same now, only one field is
necessary. There is one thing that I block to prevent wrong usage (and
guarantee a return type, even if not formalized) and for that I did add
a boolean field. There would have been other ways to do this, but this
is what I felt most comfortable with.