[#145] Add type checking to the ReferenceType
Six tests are failing (actually 3 tests, that are run once for MySQL and
once for SQLite). This is because the Id type does not pass the type
checking. I am planning a change to this, but I believe that the proper
solution is to refactor the whole Id concept out of existence, which is
a pretty big change which deserves its own commit.