[#145] Remove Id class
Instead, it's a property on the storable now that serves the same
purpose (and is in fact instantiated in the same way).

For the most part this is just refactoring, but I did introduce a type
check previously which caused some tests to fail that work again. This
also prepares for what could be an improvement in the API that I made
issue #148 for (replacing fetch by resolve).