[#137] Make unresolved references into resolvables
Also sets a option on the MySQL connection to make sure that int and
floats are returned as their respective types instead of strings. This
was different between MySQL and SQLite and this looks like a good way to
solve it, as it may also prevent other headaches in the future.

Some tests are still failing because the reference is always of the base
type, and it should be of registered child classes. I'll get right on
that.