[#56] Basic JIT datatype compiling implementation
Not entirely happy with this yet, but it's a good first step.

The tests have become a bit brittle at the moment, but this is related
to the way I am generating the input files, which is something that I'm
planning to change anyway, and then all this brittleness should
disappear.