[#137] Allow resolving unresolved collections
I don't like this implementation for a number of reasons, but it
works for the most part, which is honestly a large step forward...

The one part that doesn't work yet is that the resolver is ignored other
than some type checking. This is because it would have entailed major
changes and I didn't want to make those because it would be different
with a better solution anyway.

The things I don't like are:

- The defect described above

- Extra data is fetched from the database (other properties of the
  storable the collection belongs to)

- Using a variable to call a method and set property in the way that I'm
  doing is something I really don't like to do.

A better solution would be to allow fetching of just the collection, not
the entire storable it's a part of. But that will definitely be more
work. which would solve the above issues, but there is more that it
doesn't solve:

- We're writing a Collection to each property, only to overwrite it with
  a ResolvableCollection

- GoodManners is incompatible with other modules that also extend
  Collection, which means it's taking a role which is making it more
  than a userland (or other) module, which is not my vision for the
  modules

- If I add notifications to the class to let IDEs know about the
  properties (which I plan to do at some point), there's simply no way
  to let the IDE know that the Storables will return
  ResolvableCollections instead of normal Collections. And any simple
  solution will run into the same problem as above: it can only be done
  by one module, which would raise GoodManners above its peers in a way
  that isn't congruent with my vision for it.