[#131] Add isExplicitlyResolved method
At the moment, it's still equivalent to isResolved (and is in fact a
call to that function). However, this method will stay predictable,
while a different Storage, or cached objects (including ones with
circular references) will give a more stable / predictable results.

isResolved should only be used when you need to know whether you need to
resolve something or if you want to do an action only if it doesn't
incur the overhead of retrieving it from the database.

isExplicitlyResolved should be used when your actual result depends on
it. An example would be to determine what to include in a REST json
result.