[#131] Move isExplicitlyResolved to referencer
It was a method on storables and collections, now it is one method per
member on the storable it is referenced by.

this is because we need to be able to distinguish between explicitly
resolved and unresolved nulls. The test is still failing because the
problem has not been resolved yet, but this lays the groundwork required
for solving the problem.