[#131] Fix resolved checking checking
The variable was overwritten at one point instead of concatenated to.
This had simply been broken but not tested at that point.

I also reorder a statement that was necessary to work properly given
the checking to work properly, which was hidden by it not working
correctly.

I ran into this while I was trying to add something to this function in
order to make toArray skip unresolved properties.