[#131] Fix the resolved / unresolved null problem
I went back to ->isExplicitlyResolved() for non-null values. Otherwise,
it would not be possible to handle something that is an unresolved
reference at first, but is resolved (which is explicit) later. But the
main way is is is[Property]ExplicitlyResolved() which is then answered
directly for nulls, and forwarded for other values.