[#137] Better query for resolving collections
This makes it so that the resolver is actually used, which was the main
defect in my previous implementation.

It also prevents extra data being fetched and using variable for
function names, neither of which I liked.

This does not solve some other things I dislike, like the fact we're
overwriting the Collection properties right after setting them, and the
way GoodManners is taking a priviliged position while I want it to be
just a peer.