[#171] Make test less brittle
The test was relying on an order that was not guaranteed. I added a bit
so that the order is now guaranteed.

This does not seem to be failing at the moment, but it was at one point
during my testing. And that was fair, because the order was in fact not
guaranteed...