[#176] Fix clearing of collections
I was using a loop for some reason, which is strange to begin with, but
the loop that I was using was also failing beause you were editing the
Set while looping over it, and hat meant that you wouldn't go over all
elements. Now, I'm just using the simple clear method on the Set.